# openapi.yaml
openapi: 3.0.3
info:
  title: AptBooks Backend API
  version: "1.0.0"
  description: Phase 1 API (Foundation + Accounting Kernel)
servers:
  - url: http://localhost:4000
    description: Local
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            missingToken:
              value: { error: "Missing bearer token" }
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            forbidden:
              value: { error: "Forbidden" }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            notFound:
              value: { error: "Not found" }
    Conflict:
      description: Conflict
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            conflict:
              value: { error: "Conflict" }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    BusinessPartner:
      type: object
      properties:
        id: { type: string, format: uuid }
        organization_id: { type: string, format: uuid }
        type: { type: string, enum: [customer, vendor] }
        code: { type: string, example: "CUST-0001" }
        name: { type: string, example: "Acme Trading Ltd" }
        status: { type: string, enum: [active, inactive], example: "active" }
        default_receivable_account_id: { type: string, format: uuid, nullable: true }
        default_payable_account_id: { type: string, format: uuid, nullable: true }
        payment_term_id: { type: string, format: uuid, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateBusinessPartnerRequest:
      type: object
      required: [type, code, name]
      properties:
        type: { type: string, enum: [customer, vendor] }
        code: { type: string, example: "CUST-0001" }
        name: { type: string, example: "Acme Trading Ltd" }
        defaultReceivableAccountId: { type: string, format: uuid, nullable: true }
        defaultPayableAccountId: { type: string, format: uuid, nullable: true }
        paymentTermId: { type: string, format: uuid, nullable: true }
        status: { type: string, enum: [active, inactive], default: active }
    UpdateBusinessPartnerRequest:
      type: object
      properties:
        code: { type: string }
        name: { type: string }
        defaultReceivableAccountId: { type: string, format: uuid, nullable: true }
        defaultPayableAccountId: { type: string, format: uuid, nullable: true }
        paymentTermId: { type: string, format: uuid, nullable: true }
        status: { type: string, enum: [active, inactive] }
    PartnerContact:
      type: object
      properties:
        id: { type: string, format: uuid }
        partner_id: { type: string, format: uuid }
        name: { type: string, example: "Ama Mensah" }
        email: { type: string, example: "ama@example.com", nullable: true }
        phone: { type: string, example: "+233XXXXXXXXX", nullable: true }
        is_primary: { type: boolean, example: true }
        created_at: { type: string, format: date-time }
    UpsertPartnerContactRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, example: "Ama Mensah" }
        email: { type: string, example: "ama@example.com", nullable: true }
        phone: { type: string, example: "+233XXXXXXXXX", nullable: true }
        isPrimary: { type: boolean, default: false }
    PartnerAddress:
      type: object
      properties:
        id: { type: string, format: uuid }
        partner_id: { type: string, format: uuid }
        line1: { type: string, example: "House 12, Ring Road" }
        line2: { type: string, nullable: true }
        city: { type: string, example: "Accra", nullable: true }
        region: { type: string, example: "Greater Accra", nullable: true }
        country_code: { type: string, example: "GH", nullable: true }
        postal_code: { type: string, nullable: true }
        is_primary: { type: boolean, example: true }
        created_at: { type: string, format: date-time }
    UpsertPartnerAddressRequest:
      type: object
      required: [line1]
      properties:
        line1: { type: string, example: "House 12, Ring Road" }
        line2: { type: string, nullable: true }
        city: { type: string, nullable: true }
        region: { type: string, nullable: true }
        countryCode: { type: string, nullable: true, example: "GH" }
        postalCode: { type: string, nullable: true }
        isPrimary: { type: boolean, default: false }
...
security:
  - BearerAuth: []
paths:
...
  /modules/business/partners:
    get:
      tags: [Business Partners]
      summary: List business partners
      parameters:
        - in: query
          name: type
          schema: { type: string, enum: [customer, vendor] }
        - in: query
          name: status
          schema: { type: string, enum: [active, inactive] }
        - in: query
          name: q
          schema: { type: string, description: "Search by code/name (implementation-specific)" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/BusinessPartner" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }

    post:
      tags: [Business Partners]
      summary: Create business partner
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateBusinessPartnerRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BusinessPartner" }
        "400":
          description: Bad request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /modules/business/partners/{id}:
    get:
      tags: [Business Partners]
      summary: Get partner by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BusinessPartner" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

    patch:
      tags: [Business Partners]
      summary: Update partner
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateBusinessPartnerRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BusinessPartner" }
        "400":
          description: Bad request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /modules/business/partners/{id}/contacts:
    post:
      tags: [Business Partners]
      summary: Upsert partner contact (supports primary contact)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpsertPartnerContactRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PartnerContact" }
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PartnerContact" }
        "400":
          description: Bad request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /modules/business/partners/{id}/addresses:
    post:
      tags: [Business Partners]
      summary: Upsert partner address (supports primary address)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpsertPartnerAddressRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PartnerAddress" }
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PartnerAddress" }
        "400":
          description: Bad request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...